seperated geometry classes to seperate file
