+++ yall/python/yall/networktables/targets.py
@@ -1,8 +1,8 @@
-from wpimath import geometry
-from util import LimelightUtils
-
 from dataclasses import dataclass, field
-from typing import List, Dict, Any, Optional
+from typing import Any, Dict, List, Optional
+
+from util import LimelightUtils
+from wpimath import geometry
 
 
 @dataclass
